Rename end_of_statement to end_of_last_statement
